Completely reworked everything
Sort of some anti-progress here, but I think it'll work out better.
Instead of storing numbers in Strings, data is stored in a BigNumber
object, mainly in an array. This makes for way less processing needed
for any operations.

Anatomy of a BigNumber object:
- data: an Array that stores the meat of the number. Numbers are split
into groups of 3 base-10 digits and put in the array from least- to
most-significant. Extra zeroes at each end are removed.
BigNumber(1234.56).data is [560,234,1].
- decs: a Number that keeps track of how many decimal entries are in
this.data. If negative, this.data contains implicit zeroes at the left
end; if larger than the length of this.data, this.data contains implicit
zeroes at the right end. BigNumber(1234.56).decs is 1;
BigNumber(1000).decs is -1; BigNumber(.0001).decs is 2.
- sign: a Number that represents the sign of the number. Either -1, 0,
or 1.

Added functions:
- BigNumber(x), new BigNumber(x), or BigNumber.parse(x): returns x
converted to a BigNumber object.
- BigNumber.prototype.toString(): returns the string value of the
BigNumber object.

Currently supported BigNumber(x) formats:
- Number: 1234
- Array of digits: [1,2,3,4], [1,234], etc.
- String: "1234", "1,234", "1 234", "1_234", "1.234e3", "1.234e+3",
"1234000e-3", "-1234", "1.234"